% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{check_rstudioapi}
\alias{check_rstudioapi}
\title{Check for required RStudio API package}
\usage{
check_rstudioapi()
}
\value{
invisibly returns \code{NULL}.
}
\description{
Ensures the \code{rstudioapi} package is available, as it is required for addin functionality.
}
\keyword{internal}
