% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{format_code}
\alias{format_code}
\title{Format code}
\usage{
format_code(source_code, configuration)
}
\arguments{
\item{source_code}{(\code{character}) the R code to format}
}
\value{
(\code{character}) the formatted code
}
\description{
Format code
}
\keyword{internal}
