% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_config}
\alias{get_config}
\title{Parse the config file and return the configuration}
\usage{
get_config(path)
}
\arguments{
\item{path}{(\code{character(1)}) the file path of the configuration
file}
}
\value{
(\code{list})
}
\description{
Parse the config file and return the configuration
}
\keyword{internal}
