% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{get_printed_symbol}
\alias{get_printed_symbol}
\title{Fallback to simpler symbol if console does not support Unicode characters}
\usage{
get_printed_symbol(symbol, fallback_symbol)
}
\arguments{
\item{symbol}{(\code{character}) to print by default}

\item{fallback_symbol}{(\code{character}) to print if the option \code{tergo.unicode}
is \code{FALSE}.}
}
\description{
Fallback to simpler symbol if console does not support Unicode characters
}
\keyword{internal}
