% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{is_in_ignored_paths}
\alias{is_in_ignored_paths}
\title{Check whether a path is in ignored paths}
\usage{
is_in_ignored_paths(path, ignored_paths)
}
\value{
(\code{logical}) whether the path is in the ignored paths.
}
\description{
Check whether a path is in ignored paths
}
\keyword{internal}
