% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_active_file_addin}
\alias{style_active_file_addin}
\title{Style the active file (RStudio addin)}
\usage{
style_active_file_addin()
}
\description{
Styles the currently active file in the RStudio editor and saves the formatted code.
}
\keyword{internal}
