% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_pkg_addin}
\alias{style_pkg_addin}
\title{Style the current package (RStudio addin)}
\usage{
style_pkg_addin()
}
\value{
invisibly returns \code{NULL}.
}
\description{
Automatically styles all R code in the current project/package using
\code{\link[tergo]{style_pkg}}. If not called within a project, it
defaults to the current working directory.
}
\keyword{internal}
