% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection_addin}
\alias{style_selection_addin}
\title{Style the selected text (RStudio addin)}
\usage{
style_selection_addin()
}
\description{
This addin will retrieve the current text selection(s) in the
RStudio editor, run it through \code{style_text()}, and then replace
the selected text with the formatted version.
}
\keyword{internal}
