% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{style_text}
\alias{style_text}
\title{Style text}
\usage{
style_text(text, configuration = list())
}
\arguments{
\item{text}{(\code{character}) the text to format}

\item{configuration}{(\code{list}) the configuration for formatting}
}
\description{
This function is vectorized.
}
