% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergo.R
\docType{package}
\name{tergo-package}
\alias{tergo}
\alias{tergo-package}
\title{A library to make your source code pretty.}
\description{
Provides a set of functions that allow users for styling their R code according to the 'tidyverse' style guide. The package uses a native Rust implementation to ensure the highest performance. Learn more about 'tergo' at \url{https://rtergo.pagacz.io}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rtergo.pagacz.io}
  \item \url{https://github.com/kpagacz/tergo}
  \item Report bugs at \url{https://github.com/kpagacz/tergo/issues}
}

}
\author{
\strong{Maintainer}: Konrad Pagacz \email{konrad.pagacz@gmail.com}

Other contributors:
\itemize{
  \item Maciej Nasinski \email{nasinski.maciej@gmail.com} [contributor]
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
