% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_default_config}
\alias{get_default_config}
\title{Get the default configuration}
\usage{
get_default_config()
}
\value{
\code{list} with the default configuration
}
\description{
Get the default configuration
}
\keyword{internal}
