% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{style_file}
\alias{style_file}
\title{Style a file}
\usage{
style_file(file, configuration = list())
}
\arguments{
\item{file}{(\code{character}) the file to format}

\item{configuration}{(\code{list}) The path to the configuration for formatting. Default \code{list()}.}
}
\description{
Style a file
}
