% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{style_file}
\alias{style_file}
\title{Style a file}
\usage{
style_file(configuration = list())
}
\arguments{
\item{configuration}{(\code{list}) the configuration for formatting}
}
\description{
Style a file
}
