% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection_addin}
\alias{style_selection_addin}
\title{Style the selected text (RStudio addin)}
\usage{
style_selection_addin()
}
\description{
Styles the selected text in the RStudio editor, replacing it with the formatted version.
}
\keyword{internal}
