% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{style_text}
\alias{style_text}
\title{Style text}
\usage{
style_text(text, configuration = list())
}
\arguments{
\item{text}{(\code{character}) the text to style}

\item{configuration}{(\code{list}) Configuration for formatting. Default \code{list()}.}
}
\value{
(\code{character}) the text formatted as R code
}
\description{
Style text
}
\details{
This function is vectorized.
To see possible configuration options, see \code{\link[=get_default_config]{get_default_config()}}.
}
\examples{
code <- "function(){}"
style_text(code)

code <- c("function(){}", "A<-7")
style_text(code)
}
