% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergo.R
\docType{package}
\name{tergo-package}
\alias{tergo}
\alias{tergo-package}
\title{A library to make your source code pretty.}
\description{
`tergo` styles your R source code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kpagacz/tergo/tree/main/antidotum/tergo}
  \item Report bugs at \url{https://github.com/kpagacz/tergo/issues}
}

}
\author{
\strong{Maintainer}: Konrad Pagacz \email{konrad.pagacz@gmail.com}

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
